% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoExp-package.R
\docType{package}
\name{DoExp-package}
\alias{DoExp}
\alias{DoExp-package}
\title{DoExp: Design of Experiments with Hasse Diagrams and Anatomy Tables}
\description{
Return Hasse Diagrams or Anatomy Tables from structure formulaes.
}
\author{
\strong{Maintainer}: Leonardo Peressim \email{leonardoperessim@usp.br} (\href{https://orcid.org/0000-0002-5378-6903}{ORCID})

Authors:
\itemize{
  \item Renata Sermarini \email{ralcarde@usp.br} (\href{https://orcid.org/0000-0001-6425-9626}{ORCID})
}

}
\keyword{internal}
