% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hassediagram.R
\name{Hassediagram}
\alias{Hassediagram}
\title{Hassediagram}
\usage{
Hassediagram(
  units = NULL,
  trt = NULL,
  data = NULL,
  random = NULL,
  advancedrandom = NULL,
  leftcolor = "blue",
  rightcolor = "red"
)
}
\arguments{
\item{units}{structural formula of units. Must be enclosed in quotation marks.}

\item{trt}{structural formula of treatments. Must be enclosed in quotation marks.}

\item{data}{dataframe containing the variables in the formulas.}

\item{random}{random-effect variables that extend randomness to the factors
with respect to which they are marginal.}

\item{advancedrandom}{random-effect variables that do not extend randomness
to the factors with respect to which they are marginal.}

\item{leftcolor}{sets the color of everything to the left of the dot.
Default is "blue".}

\item{rightcolor}{sets the color of everything to the right of the dot.
Default is "red".}
}
\description{
From the structural formulas of units and treatments and a data frame,
returns the Hasse diagrams with the quantities of interest: sources of
variation, degrees of freedom, core matrices of the quadratic forms for
sums of squares, and the contributions to the expectations of mean squares.
}
