% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectHasse.R
\name{objectHasse}
\alias{objectHasse}
\title{objectHasse}
\usage{
objectHasse(
  sform = NULL,
  data = NULL,
  type = NULL,
  random = NULL,
  advancedrandom = NULL
)
}
\arguments{
\item{sform}{a structural formula as a character.}

\item{data}{data.frame containing the variables named in structural the
formulas.}

\item{type}{indicates whether the factors are unit factors or treatment factors.
For unit factors: "unit", "units", "unrandomized", or "recipient".
For treatment factors: "treatment", "trt", "randomized", or "allocated".}

\item{random}{a vector containing the variables related to the random factors.
All factors for which the stated factors are marginal will be assumed to be
random.}

\item{advancedrandom}{a vector containing the variables related to the random.
factors. Here the factor for which the stated factors are marginal will not
be assumed to be random.}
}
\description{
Returns the Hasse diagram with: source of variation, degrees of freedom,
matrices of the quadratic forms for sums of squares and the contributions to
the expectations of mean squares,from structural formulas.
}
\examples{
if (requireNamespace("dae", quietly = TRUE)) {
  blocks <- dae::fac.gen(list(
    Blocks = 4,
    A = 2,
    C = 3
  ))

  blocks$Plots <- rep(1:6, times = 4)

  units <- objectHasse(
    sform = "Blocks/Plots",
    data = blocks,
    type = "units",
    random = NULL,
    advancedrandom = NULL
  )

  treatments <- objectHasse(
    sform = "A*C",
    data = blocks,
    type = "trt",
    random = c("A"),
    advancedrandom = NULL
  )
}
}
