% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHasse.R
\name{plotHasse}
\alias{plotHasse}
\title{plotHasse}
\usage{
plotHasse(
  diagram = NULL,
  which = NULL,
  titlehd = "Hasse diagram",
  subtitlehd = "",
  leftcolor = "blue",
  rightcolor = "red"
)
}
\arguments{
\item{diagram}{an object from objectHasse with informations for ploting.}

\item{which}{a character to select which graph will be generated.
Choose between: "sv", "df", "M", and "EMS".}

\item{titlehd}{The graph title. Can be NULL. Default: "Hasse diagram".}

\item{subtitlehd}{The graph subtitle. Can be NULL. Default: NULL.}

\item{leftcolor}{the color of everything to the left of the dot in the Hasse
diagram. Default is "blue".}

\item{rightcolor}{the color of everything to the right of the dot in the
Hasse diagram. Default is "red".}
}
\description{
Returns the Hasse diagram with: source of variation, degrees of freedom,
matrices of the quadratic forms for sums of squares and the contributions to
the expectations of mean squares,from structural formulas.
}
\examples{
if (requireNamespace("dae", quietly = TRUE)) {
  blocks <- dae::fac.gen(list(
    Blocks = 4,
    A = 2,
    C = 3
  ))

  blocks$Plots <- rep(1:6, times = 4)

  units <- objectHasse(
    sform = "Blocks/Plots",
    data = blocks,
    type = "units",
    random = NULL,
    advancedrandom = NULL
  )

  treatments <- objectHasse(
    sform = "A*C",
    data = blocks,
    type = "trt",
    random = c("A"),
    advancedrandom = NULL
  )

  svunits <- plotHasse(diagram = units, which = "sv")
  svtrt   <- plotHasse(diagram = treatments, which = "sv")

svunits
svtrt
}
}
