% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hassediagramdf2.R
\name{Hassediagramdf}
\alias{Hassediagramdf}
\title{Hassediagramdf}
\usage{
Hassediagramdf(
  sform = NULL,
  data = NULL,
  rightcolor = "red",
  titlehd = "Hasse diagram",
  subtitlehd = "Levels & Degrees of freedom"
)
}
\arguments{
\item{sform}{a structural formula as a character.}

\item{data}{data.frame containing the variables named in structural the
formulas.}

\item{rightcolor}{the color of everything to the right of the dot in the
Hasse diagram. Default is "red".}

\item{titlehd}{The graph title. Can be NULL. Default: "Hasse diagram".}

\item{subtitlehd}{The graph subtitle. Can be NULL. Default: "Levels & Degrees of freedom".}
}
\description{
Returns the Hasse diagram with: source of variation, degrees of freedom,
matrices of the quadratic forms for sums of squares and the contributions to
the expectations of mean squares,from structural formulas.
}
\examples{
if (requireNamespace("dae", quietly = TRUE)) {
  blocks <- dae::fac.gen(list(
    Blocks = 4,
    A = 2,
    C = 3
  ))

  blocks$Plots <- rep(1:6, times = 4)

  unitsdf <- Hassediagramdf(
    sform = "Blocks/Plots",
    data = blocks,
    rightcolor = "red",
    titlehd = "Hasse diagram",
    subtitlehd = "Levels & Degrees of freedom"
  )

  trtdf <- Hassediagramdf(
    sform = "A*C",
    data = blocks,
    rightcolor = "blue",
    titlehd = "Hasse diagram",
    subtitlehd = "Levels & Degrees of freedom"
  )

  unitsdf
  trtdf
}
}
