% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toa.R
\name{toa}
\alias{toa}
\title{toa}
\usage{
toa(units, trt, data)
}
\arguments{
\item{units}{structural formula of units. Must be enclosed in quotation marks.}

\item{trt}{structural formula of treatments. Must be enclosed in quotation marks.}

\item{data}{dataframe containing the variables in the formulas.}
}
\description{
From the structural formulas of units and treatments and a data frame,
returns the Anatomy Table with degrees of freedom and the a-effiency value.
}
